+++ blockfrost-client/CHANGELOG.md
@@ -5,6 +5,7 @@
   * `listScripts` for `/scripts` endpoint, principled `listScripts'`
   * `getScript` for `/scripts/${script_hash}` endpoint
   * `getScriptRedeemers` for `/scripts/${script_hash}/redeemers` endpoint, principled `getScriptRedeemers'`
+  * `getAddressUtxosAsset` for `/addresses/${address}/utxos/${asset}`, principled `getAddressUtxosAsset'`
 
 # Version [0.1.0.0](https://github.com/blockfrost/blockfrost-haskell/compare/initial...0.1.0.0) (2021-09-14)